FU r102073: * (bug 32367) Fixed SpecialMovePage to not call wfLocalFile() on a non file title and expect an actual File back. Previously "worked" due to an old file title checking loophole (File objects with non File: titles). * Use File::normalizeTitle() in some more functions that were still doing their own incomplete normalization. * Updated FileRepo::newFile() docs to reflect that it can return null; wfLocalFile() docs already mentioned this.

+++ includes/specials/SpecialMovepage.php
@@ -376,9 +376,11 @@ class MovePageForm extends UnlistedSpecialPage {
 			$reason = wfMessage( 'delete_and_move_reason', $ot )->inContentLanguage()->text();
 
 			// Delete an associated image if there is
-			$file = wfLocalFile( $nt );
-			if( $file->exists() ) {
-				$file->delete( $reason, false );
+			if ( $nt->getNamespace() == NS_FILE ) {
+				$file = wfLocalFile( $nt );
+				if ( $file->exists() ) {
+					$file->delete( $reason, false );
+				}
 			}
 
 			$error = ''; // passed by ref
